+++ t/resolver.t
@@ -2,13 +2,10 @@ use lib 't';
 use TestAPIcast 'no_plan';
 
 $ENV{TEST_NGINX_HTTP_CONFIG} = "$TestAPIcast::path/http.d/*.conf";
-$ENV{RESOLVER} = '127.0.1.1:5353';
+$ENV{TEST_NGINX_RESOLVER} = '127.0.1.1:5353';
 
 $ENV{TEST_NGINX_RESOLV_CONF} = "$Test::Nginx::Util::HtmlDir/resolv.conf";
 
-env_to_nginx(
-    'RESOLVER'
-);
 master_on();
 log_level('warn');
 run_tests();
@@ -17,9 +14,11 @@ __DATA__
 
 === TEST 1: uses all resolvers
 both RESOLVER env variable and resolvers in resolv.conf should be used
+--- main_config
+env RESOLVER=$TEST_NGINX_RESOLVER;
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
-  init_by_lua_block {
+  init_worker_by_lua_block {
     require('resty.resolver').init('$TEST_NGINX_RESOLV_CONF')
   }
 --- config
@@ -67,3 +66,27 @@ servers: 2
 2.3.4.5:6789
 --- no_error_log
 [error]
+
+
+=== TEST 3: can have ipv6 RESOLVER
+RESOLVER env variable can be IPv6 address
+--- main_config
+env RESOLVER=[dead::beef]:5353;
+--- http_config
+  lua_package_path "$TEST_NGINX_LUA_PATH";
+  init_worker_by_lua_block {
+    require('resty.resolver').init('$TEST_NGINX_RESOLV_CONF')
+  }
+--- config
+  location = /t {
+    content_by_lua_block {
+      local nameservers = require('resty.resolver').nameservers()
+      ngx.say('nameservers: ', #nameservers, ' ', tostring(nameservers[1]))
+    }
+  }
+--- request
+GET /t
+--- response_body
+nameservers: 1 [dead::beef]:5353
+--- user_files
+>>> resolv.conf